Move config migration logic to "server-config" module
